// src/components/workload/research/research-workload.interface.ts
import IResearchActivity from '../../activity/research/research-activity.interface';
import { Document } from 'mongoose';

export interface IResearchWorkloadPerActivity {
  activity: IResearchActivity;
  totalHoursPerActivity: number;
  percentageOfWorkFocusPerActivity: number;
  percentageOfAnnualHoursPerActivity: number;
  percentageOfTotalHoursPerActivity: number;
}

export default interface IResearchWorkload extends Document {
  userId: string;
  researchWorkloads: IResearchWorkloadPerActivity[];
  globalTarrif: number;
  totalHoursPerUser: number;
  percentageOfWorkFocusPerUser: number;
  percentageOfAnnualHoursPerUser: number;
  percentageOfTotalHoursPerUser: number;
}